refactor: simplify turbidity overflow guard condition
Agent-Logs-Url: https://github.com/suikaxxc/suika_ovo/sessions/959fb3e9-4a3e-4a58-b033-8b22b0282de4

+++ code-1.0/applications/sample/wifi-iot/app/code_v2.0/suika_demo/turbidity_sensor.c
@@ -113,7 +113,7 @@ static int CalculateNTUFromRaw(unsigned short raw)
         // to avoid unbounded span inflation under long-term clear-water sampling.
         if (g_raw_turbid_ref < g_raw_clear_ref) {
             unsigned short delta = (unsigned short)(g_raw_clear_ref - prevClearRef);
-            if ((unsigned int)delta <= ((unsigned int)USHRT_MAX - (unsigned int)g_raw_turbid_ref)) {
+            if (delta <= (USHRT_MAX - g_raw_turbid_ref)) {
                 unsigned short shifted = (unsigned short)(g_raw_turbid_ref + delta);
                 if (shifted < g_raw_clear_ref) {
                     g_raw_turbid_ref = shifted;
